Refactoring
- Move functions into currency.py
