+++ src/backend/InvenTree/common/currency.py
@@ -1,12 +1,16 @@
 """Helper functions for currency support."""
 
+import decimal
 import logging
+import math
 
 from django.core.cache import cache
 from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 
 from moneyed import CURRENCIES
+
+import InvenTree.helpers
 
 logger = logging.getLogger('inventree')
 
@@ -68,6 +72,11 @@ def currency_codes() -> list:
     return list(valid_codes)
 
 
+def currency_codes_default_list() -> str:
+    """Return a comma-separated list of default currency codes."""
+    return 'AUD,CAD,CNY,EUR,GBP,JPY,NZD,USD'
+
+
 def currency_code_mappings():
     """Returns the current currency choices."""
     return [(a, CURRENCIES[a].name) for a in currency_codes()]
@@ -115,3 +124,91 @@ def validate_currency_codes(value):
             raise ValidationError(_('Invalid currency code') + f": '{code}'")
 
     return list(valid_currencies)
+
+
+def currency_exchange_plugins() -> list:
+    """Return a list of plugin choices which can be used for currency exchange."""
+    try:
+        from plugin import registry
+
+        plugs = registry.with_mixin('currencyexchange', active=True)
+    except Exception:
+        plugs = []
+
+    return [('', _('No plugin'))] + [(plug.slug, plug.human_name) for plug in plugs]
+
+
+def get_price(
+    instance,
+    quantity,
+    moq=True,
+    multiples=True,
+    currency=None,
+    break_name: str = 'price_breaks',
+):
+    """Calculate the price based on quantity price breaks.
+
+    - Don't forget to add in flat-fee cost (base_cost field)
+    - If MOQ (minimum order quantity) is required, bump quantity
+    - If order multiples are to be observed, then we need to calculate based on that, too
+    """
+    from common.currency import currency_code_default
+
+    if hasattr(instance, break_name):
+        price_breaks = getattr(instance, break_name).all()
+    else:
+        price_breaks = []
+
+    # No price break information available?
+    if len(price_breaks) == 0:
+        return None
+
+    # Check if quantity is fraction and disable multiples
+    multiples = quantity % 1 == 0
+
+    # Order multiples
+    if multiples:
+        quantity = int(math.ceil(quantity / instance.multiple) * instance.multiple)
+
+    pb_found = False
+    pb_quantity = -1
+    pb_cost = 0.0
+
+    if currency is None:
+        # Default currency selection
+        currency = currency_code_default()
+
+    pb_min = None
+    for pb in price_breaks:
+        # Store smallest price break
+        if not pb_min:
+            pb_min = pb
+
+        # Ignore this pricebreak (quantity is too high)
+        if pb.quantity > quantity:
+            continue
+
+        pb_found = True
+
+        # If this price-break quantity is the largest so far, use it!
+        if pb.quantity > pb_quantity:
+            pb_quantity = pb.quantity
+
+            # Convert everything to the selected currency
+            pb_cost = pb.convert_to(currency)
+
+    # Use smallest price break
+    if not pb_found and pb_min:
+        # Update price break information
+        pb_quantity = pb_min.quantity
+        pb_cost = pb_min.convert_to(currency)
+        # Trigger cost calculation using smallest price break
+        pb_found = True
+
+    # Convert quantity to decimal.Decimal format
+    quantity = decimal.Decimal(f'{quantity}')
+
+    if pb_found:
+        cost = pb_cost * quantity
+        return InvenTree.helpers.normalize(cost + instance.base_cost)
+    return None
